refactor: replace LightTable with SimpleTable
